Initialize ENI ack timer if needed upon restart.
Previously the ack timer of each ENI attachment is not initialized upon agent restarts. If agent is stopped and started after an ENI attachment is added to state but before it's acked, we will end up with one of these: (1) If the attachment has not expired, the agent crashes on NPE while attempting tostop the ack timer after acking the attachment; (2) If the attachment has expired, the ENI attachment is left in agent state (since the task will not start in that case, we won't be removing it during task cleanup). This commit fixes the two situations by attempting to initialize the ack timer upon restart, so that - (1) if the attachment has not expired, the ack timer will be initialized so that we won't get NPE when acking it; (2) if the attachment has expired, it gets removed from agent state.